requirements.txt
# Cross-Store Analysis System Requirements

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Performance
joblib>=1.3.0

# Date/Time
python-dateutil>=2.8.0
